+++ packages/controllers/ajax_new_key.go
@@ -137,7 +137,7 @@ func (c *Controller) AjaxNewKey() interface{} {
 	}
 	var flags uint8
 
-	ctime := time.Now().Unix()
+	ctime := uint32(time.Now().Unix())
 	info := (*contract).Block.Info.(*script.ContractInfo)
 	forsign := fmt.Sprintf("%d,%d,%d,%d,%d", info.ID, ctime, uint64(idkey), stateID, flags)
 	pubhex := hex.EncodeToString(pub)
